Fixed 4

+++ app/src/main/java/d_EntreBonsGro/E_FB_SupplierCredit.kt
@@ -33,7 +33,7 @@ import java.util.Locale
 
 
 fun updateSupplierCredit(
-    supplierId: Int,
+    supplierId: Long,
     supplierTotal: Double,
     supplierPayment: Double,
     ancienCredit: Double
